@@ -1,72 +1,35 @@
 {
-  "_from": "body-parser@1.18.3",
-  "_id": "body-parser@1.18.3",
-  "_inBundle": false,
-  "_integrity": "sha1-WykhmP/dVTs6DyDe0FkrlWlVyLQ=",
-  "_location": "/body-parser",
-  "_phantomChildren": {},
-  "_requested": {
-    "type": "version",
-    "registry": true,
-    "raw": "body-parser@1.18.3",
-    "name": "body-parser",
-    "escapedName": "body-parser",
-    "rawSpec": "1.18.3",
-    "saveSpec": null,
-    "fetchSpec": "1.18.3"
-  },
-  "_requiredBy": [
-    "/express"
-  ],
-  "_resolved": "https://registry.npmjs.org/body-parser/-/body-parser-1.18.3.tgz",
-  "_shasum": "5b292198ffdd553b3a0f20ded0592b956955c8b4",
-  "_spec": "body-parser@1.18.3",
-  "_where": "/Users/didin/NodeApps/node-graphql/node_modules/express",
-  "bugs": {
-    "url": "https://github.com/expressjs/body-parser/issues"
-  },
-  "bundleDependencies": false,
-  "contributors": [
-    {
-      "name": "Douglas Christopher Wilson",
-      "email": "doug@somethingdoug.com"
-    },
-    {
-      "name": "Jonathan Ong",
-      "email": "me@jongleberry.com",
-      "url": "http://jongleberry.com"
-    }
-  ],
-  "dependencies": {
-    "bytes": "3.0.0",
-    "content-type": "~1.0.4",
-    "debug": "2.6.9",
-    "depd": "~1.1.2",
-    "http-errors": "~1.6.3",
-    "iconv-lite": "0.4.23",
-    "on-finished": "~2.3.0",
-    "qs": "6.5.2",
-    "raw-body": "2.3.3",
-    "type-is": "~1.6.16"
-  },
-  "deprecated": false,
+  "name": "body-parser",
   "description": "Node.js body parsing middleware",
-  "devDependencies": {
-    "eslint": "4.19.1",
-    "eslint-config-standard": "11.0.0",
-    "eslint-plugin-import": "2.11.0",
-    "eslint-plugin-markdown": "1.0.0-beta.6",
-    "eslint-plugin-node": "6.0.1",
-    "eslint-plugin-promise": "3.7.0",
-    "eslint-plugin-standard": "3.1.0",
-    "istanbul": "0.4.5",
-    "methods": "1.1.2",
-    "mocha": "2.5.3",
-    "safe-buffer": "5.1.2",
-    "supertest": "1.1.0"
+  "version": "2.2.0",
+  "contributors": [
+    "Douglas Christopher Wilson <doug@somethingdoug.com>",
+    "Jonathan Ong <me@jongleberry.com> (http://jongleberry.com)"
+  ],
+  "license": "MIT",
+  "repository": "expressjs/body-parser",
+  "dependencies": {
+    "bytes": "^3.1.2",
+    "content-type": "^1.0.5",
+    "debug": "^4.4.0",
+    "http-errors": "^2.0.0",
+    "iconv-lite": "^0.6.3",
+    "on-finished": "^2.4.1",
+    "qs": "^6.14.0",
+    "raw-body": "^3.0.0",
+    "type-is": "^2.0.0"
   },
-  "engines": {
-    "node": ">= 0.8"
+  "devDependencies": {
+    "eslint": "8.34.0",
+    "eslint-config-standard": "14.1.1",
+    "eslint-plugin-import": "2.27.5",
+    "eslint-plugin-markdown": "3.0.0",
+    "eslint-plugin-node": "11.1.0",
+    "eslint-plugin-promise": "6.1.1",
+    "eslint-plugin-standard": "4.1.0",
+    "mocha": "^11.1.0",
+    "nyc": "^17.1.0",
+    "supertest": "^7.0.0"
   },
   "files": [
     "lib/",
@@ -74,18 +37,13 @@
     "HISTORY.md",
     "index.js"
   ],
-  "homepage": "https://github.com/expressjs/body-parser#readme",
-  "license": "MIT",
-  "name": "body-parser",
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/expressjs/body-parser.git"
+  "engines": {
+    "node": ">=18"
   },
   "scripts": {
-    "lint": "eslint --plugin markdown --ext js,md .",
-    "test": "mocha --require test/support/env --reporter spec --check-leaks --bail test/",
-    "test-cov": "istanbul cover node_modules/mocha/bin/_mocha -- --require test/support/env --reporter dot --check-leaks test/",
-    "test-travis": "istanbul cover node_modules/mocha/bin/_mocha --report lcovonly -- --require test/support/env --reporter spec --check-leaks test/"
-  },
-  "version": "1.18.3"
+    "lint": "eslint .",
+    "test": "mocha --reporter spec --check-leaks test/",
+    "test-ci": "nyc --reporter=lcovonly --reporter=text npm test",
+    "test-cov": "nyc --reporter=html --reporter=text npm test"
+  }
 }
