@@ -1,49 +1,30 @@
 {
-  "_from": "content-disposition@0.5.2",
-  "_id": "content-disposition@0.5.2",
-  "_inBundle": false,
-  "_integrity": "sha1-DPaLud318r55YcOoUXjLhdunjLQ=",
-  "_location": "/content-disposition",
-  "_phantomChildren": {},
-  "_requested": {
-    "type": "version",
-    "registry": true,
-    "raw": "content-disposition@0.5.2",
-    "name": "content-disposition",
-    "escapedName": "content-disposition",
-    "rawSpec": "0.5.2",
-    "saveSpec": null,
-    "fetchSpec": "0.5.2"
-  },
-  "_requiredBy": [
-    "/express"
-  ],
-  "_resolved": "https://registry.npmjs.org/content-disposition/-/content-disposition-0.5.2.tgz",
-  "_shasum": "0cf68bb9ddf5f2be7961c3a85178cb85dba78cb4",
-  "_spec": "content-disposition@0.5.2",
-  "_where": "/Users/didin/NodeApps/node-graphql/node_modules/express",
-  "bugs": {
-    "url": "https://github.com/jshttp/content-disposition/issues"
-  },
-  "bundleDependencies": false,
-  "contributors": [
-    {
-      "name": "Douglas Christopher Wilson",
-      "email": "doug@somethingdoug.com"
-    }
-  ],
-  "deprecated": false,
+  "name": "content-disposition",
   "description": "Create and parse Content-Disposition header",
-  "devDependencies": {
-    "eslint": "3.11.1",
-    "eslint-config-standard": "6.2.1",
-    "eslint-plugin-promise": "3.3.0",
-    "eslint-plugin-standard": "2.0.1",
-    "istanbul": "0.4.5",
-    "mocha": "1.21.5"
+  "version": "1.0.0",
+  "author": "Douglas Christopher Wilson <doug@somethingdoug.com>",
+  "license": "MIT",
+  "keywords": [
+    "content-disposition",
+    "http",
+    "rfc6266",
+    "res"
+  ],
+  "repository": "jshttp/content-disposition",
+  "dependencies": {
+    "safe-buffer": "5.2.1"
   },
-  "engines": {
-    "node": ">= 0.6"
+  "devDependencies": {
+    "deep-equal": "1.0.1",
+    "eslint": "7.32.0",
+    "eslint-config-standard": "13.0.1",
+    "eslint-plugin-import": "2.25.3",
+    "eslint-plugin-markdown": "2.2.1",
+    "eslint-plugin-node": "11.1.0",
+    "eslint-plugin-promise": "5.2.0",
+    "eslint-plugin-standard": "4.1.0",
+    "mocha": "^9.2.2",
+    "nyc": "15.1.0"
   },
   "files": [
     "LICENSE",
@@ -51,24 +32,13 @@
     "README.md",
     "index.js"
   ],
-  "homepage": "https://github.com/jshttp/content-disposition#readme",
-  "keywords": [
-    "content-disposition",
-    "http",
-    "rfc6266",
-    "res"
-  ],
-  "license": "MIT",
-  "name": "content-disposition",
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/jshttp/content-disposition.git"
+  "engines": {
+    "node": ">= 0.6"
   },
   "scripts": {
     "lint": "eslint .",
     "test": "mocha --reporter spec --bail --check-leaks test/",
-    "test-cov": "istanbul cover node_modules/mocha/bin/_mocha -- --reporter dot --check-leaks test/",
-    "test-travis": "istanbul cover node_modules/mocha/bin/_mocha --report lcovonly -- --reporter spec --check-leaks test/"
-  },
-  "version": "0.5.2"
+    "test-ci": "nyc --reporter=lcovonly --reporter=text npm test",
+    "test-cov": "nyc --reporter=html --reporter=text npm test"
+  }
 }
