@@ -1,55 +1,29 @@
 {
-  "_from": "depd@~1.1.2",
-  "_id": "depd@1.1.2",
-  "_inBundle": false,
-  "_integrity": "sha1-m81S4UwJd2PnSbJ0xDRu0uVgtak=",
-  "_location": "/depd",
-  "_phantomChildren": {},
-  "_requested": {
-    "type": "range",
-    "registry": true,
-    "raw": "depd@~1.1.2",
-    "name": "depd",
-    "escapedName": "depd",
-    "rawSpec": "~1.1.2",
-    "saveSpec": null,
-    "fetchSpec": "~1.1.2"
-  },
-  "_requiredBy": [
-    "/body-parser",
-    "/express",
-    "/http-errors",
-    "/morgan",
-    "/send"
-  ],
-  "_resolved": "https://registry.npmjs.org/depd/-/depd-1.1.2.tgz",
-  "_shasum": "9bcd52e14c097763e749b274c4346ed2e560b5a9",
-  "_spec": "depd@~1.1.2",
-  "_where": "/Users/didin/NodeApps/node-graphql/node_modules/express",
-  "author": {
-    "name": "Douglas Christopher Wilson",
-    "email": "doug@somethingdoug.com"
-  },
-  "browser": "lib/browser/index.js",
-  "bugs": {
-    "url": "https://github.com/dougwilson/nodejs-depd/issues"
-  },
-  "bundleDependencies": false,
-  "deprecated": false,
+  "name": "depd",
   "description": "Deprecate all the things",
+  "version": "2.0.0",
+  "author": "Douglas Christopher Wilson <doug@somethingdoug.com>",
+  "license": "MIT",
+  "keywords": [
+    "deprecate",
+    "deprecated"
+  ],
+  "repository": "dougwilson/nodejs-depd",
+  "browser": "lib/browser/index.js",
   "devDependencies": {
-    "beautify-benchmark": "0.2.4",
     "benchmark": "2.1.4",
-    "eslint": "3.19.0",
-    "eslint-config-standard": "7.1.0",
+    "beautify-benchmark": "0.2.4",
+    "eslint": "5.7.0",
+    "eslint-config-standard": "12.0.0",
+    "eslint-plugin-import": "2.14.0",
     "eslint-plugin-markdown": "1.0.0-beta.7",
-    "eslint-plugin-promise": "3.6.0",
-    "eslint-plugin-standard": "3.0.1",
+    "eslint-plugin-node": "7.0.1",
+    "eslint-plugin-promise": "4.0.1",
+    "eslint-plugin-standard": "4.0.0",
     "istanbul": "0.4.5",
-    "mocha": "~1.21.5"
-  },
-  "engines": {
-    "node": ">= 0.6"
+    "mocha": "5.2.0",
+    "safe-buffer": "5.1.2",
+    "uid-safe": "2.1.5"
   },
   "files": [
     "lib/",
@@ -58,23 +32,14 @@
     "index.js",
     "Readme.md"
   ],
-  "homepage": "https://github.com/dougwilson/nodejs-depd#readme",
-  "keywords": [
-    "deprecate",
-    "deprecated"
-  ],
-  "license": "MIT",
-  "name": "depd",
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/dougwilson/nodejs-depd.git"
+  "engines": {
+    "node": ">= 0.8"
   },
   "scripts": {
     "bench": "node benchmark/index.js",
     "lint": "eslint --plugin markdown --ext js,md .",
     "test": "mocha --reporter spec --bail test/",
-    "test-ci": "istanbul cover node_modules/mocha/bin/_mocha --report lcovonly -- --reporter spec --no-exit test/",
-    "test-cov": "istanbul cover node_modules/mocha/bin/_mocha -- --reporter dot test/"
-  },
-  "version": "1.1.2"
+    "test-ci": "istanbul cover --print=none node_modules/mocha/bin/_mocha -- --reporter spec test/ && istanbul report lcovonly text-summary",
+    "test-cov": "istanbul cover --print=none node_modules/mocha/bin/_mocha -- --reporter dot test/ && istanbul report lcov text-summary"
+  }
 }
