@@ -1,54 +1,26 @@
 {
-  "_from": "encodeurl@~1.0.2",
-  "_id": "encodeurl@1.0.2",
-  "_inBundle": false,
-  "_integrity": "sha1-rT/0yG7C0CkyL1oCw6mmBslbP1k=",
-  "_location": "/encodeurl",
-  "_phantomChildren": {},
-  "_requested": {
-    "type": "range",
-    "registry": true,
-    "raw": "encodeurl@~1.0.2",
-    "name": "encodeurl",
-    "escapedName": "encodeurl",
-    "rawSpec": "~1.0.2",
-    "saveSpec": null,
-    "fetchSpec": "~1.0.2"
-  },
-  "_requiredBy": [
-    "/express",
-    "/finalhandler",
-    "/send",
-    "/serve-static"
-  ],
-  "_resolved": "https://registry.npmjs.org/encodeurl/-/encodeurl-1.0.2.tgz",
-  "_shasum": "ad3ff4c86ec2d029322f5a02c3a9a606c95b3f59",
-  "_spec": "encodeurl@~1.0.2",
-  "_where": "/Users/didin/NodeApps/node-graphql/node_modules/express",
-  "bugs": {
-    "url": "https://github.com/pillarjs/encodeurl/issues"
-  },
-  "bundleDependencies": false,
-  "contributors": [
-    {
-      "name": "Douglas Christopher Wilson",
-      "email": "doug@somethingdoug.com"
-    }
-  ],
-  "deprecated": false,
+  "name": "encodeurl",
   "description": "Encode a URL to a percent-encoded form, excluding already-encoded sequences",
+  "version": "2.0.0",
+  "contributors": [
+    "Douglas Christopher Wilson <doug@somethingdoug.com>"
+  ],
+  "license": "MIT",
+  "keywords": [
+    "encode",
+    "encodeurl",
+    "url"
+  ],
+  "repository": "pillarjs/encodeurl",
   "devDependencies": {
-    "eslint": "3.19.0",
-    "eslint-config-standard": "10.2.1",
-    "eslint-plugin-import": "2.8.0",
-    "eslint-plugin-node": "5.2.1",
-    "eslint-plugin-promise": "3.6.0",
-    "eslint-plugin-standard": "3.0.1",
+    "eslint": "5.11.1",
+    "eslint-config-standard": "12.0.0",
+    "eslint-plugin-import": "2.14.0",
+    "eslint-plugin-node": "7.0.1",
+    "eslint-plugin-promise": "4.0.1",
+    "eslint-plugin-standard": "4.0.0",
     "istanbul": "0.4.5",
     "mocha": "2.5.3"
-  },
-  "engines": {
-    "node": ">= 0.8"
   },
   "files": [
     "LICENSE",
@@ -56,23 +28,13 @@
     "README.md",
     "index.js"
   ],
-  "homepage": "https://github.com/pillarjs/encodeurl#readme",
-  "keywords": [
-    "encode",
-    "encodeurl",
-    "url"
-  ],
-  "license": "MIT",
-  "name": "encodeurl",
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/pillarjs/encodeurl.git"
+  "engines": {
+    "node": ">= 0.8"
   },
   "scripts": {
     "lint": "eslint .",
     "test": "mocha --reporter spec --bail --check-leaks test/",
     "test-cov": "istanbul cover node_modules/mocha/bin/_mocha -- --reporter dot --check-leaks test/",
     "test-travis": "istanbul cover node_modules/mocha/bin/_mocha --report lcovonly -- --reporter spec --check-leaks test/"
-  },
-  "version": "1.0.2"
+  }
 }
