@@ -1,122 +1,85 @@
 {
-  "_from": "express@~4.16.0",
-  "_id": "express@4.16.4",
-  "_inBundle": false,
-  "_integrity": "sha512-j12Uuyb4FMrd/qQAm6uCHAkPtO8FDTRJZBDd5D2KOL2eLaz1yUNdUB/NOIyq0iU4q4cFarsUCrnFDPBcnksuOg==",
-  "_location": "/express",
-  "_phantomChildren": {},
-  "_requested": {
-    "type": "range",
-    "registry": true,
-    "raw": "express@~4.16.0",
-    "name": "express",
-    "escapedName": "express",
-    "rawSpec": "~4.16.0",
-    "saveSpec": null,
-    "fetchSpec": "~4.16.0"
-  },
-  "_requiredBy": [
-    "#USER",
-    "/"
-  ],
-  "_resolved": "https://registry.npmjs.org/express/-/express-4.16.4.tgz",
-  "_shasum": "fddef61926109e24c515ea97fd2f1bdbf62df12e",
-  "_spec": "express@~4.16.0",
-  "_where": "/Users/didin/NodeApps/node-graphql",
-  "author": {
-    "name": "TJ Holowaychuk",
-    "email": "tj@vision-media.ca"
-  },
-  "bugs": {
-    "url": "https://github.com/expressjs/express/issues"
-  },
-  "bundleDependencies": false,
+  "name": "express",
+  "description": "Fast, unopinionated, minimalist web framework",
+  "version": "5.1.0",
+  "author": "TJ Holowaychuk <tj@vision-media.ca>",
   "contributors": [
-    {
-      "name": "Aaron Heckmann",
-      "email": "aaron.heckmann+github@gmail.com"
-    },
-    {
-      "name": "Ciaran Jessup",
-      "email": "ciaranj@gmail.com"
-    },
-    {
-      "name": "Douglas Christopher Wilson",
-      "email": "doug@somethingdoug.com"
-    },
-    {
-      "name": "Guillermo Rauch",
-      "email": "rauchg@gmail.com"
-    },
-    {
-      "name": "Jonathan Ong",
-      "email": "me@jongleberry.com"
-    },
-    {
-      "name": "Roman Shtylman",
-      "email": "shtylman+expressjs@gmail.com"
-    },
-    {
-      "name": "Young Jae Sim",
-      "email": "hanul@hanul.me"
-    }
+    "Aaron Heckmann <aaron.heckmann+github@gmail.com>",
+    "Ciaran Jessup <ciaranj@gmail.com>",
+    "Douglas Christopher Wilson <doug@somethingdoug.com>",
+    "Guillermo Rauch <rauchg@gmail.com>",
+    "Jonathan Ong <me@jongleberry.com>",
+    "Roman Shtylman <shtylman+expressjs@gmail.com>",
+    "Young Jae Sim <hanul@hanul.me>"
+  ],
+  "license": "MIT",
+  "repository": "expressjs/express",
+  "homepage": "https://expressjs.com/",
+  "funding": {
+    "type": "opencollective",
+    "url": "https://opencollective.com/express"
+  },
+  "keywords": [
+    "express",
+    "framework",
+    "sinatra",
+    "web",
+    "http",
+    "rest",
+    "restful",
+    "router",
+    "app",
+    "api"
   ],
   "dependencies": {
-    "accepts": "~1.3.5",
-    "array-flatten": "1.1.1",
-    "body-parser": "1.18.3",
-    "content-disposition": "0.5.2",
-    "content-type": "~1.0.4",
-    "cookie": "0.3.1",
-    "cookie-signature": "1.0.6",
-    "debug": "2.6.9",
-    "depd": "~1.1.2",
-    "encodeurl": "~1.0.2",
-    "escape-html": "~1.0.3",
-    "etag": "~1.8.1",
-    "finalhandler": "1.1.1",
-    "fresh": "0.5.2",
-    "merge-descriptors": "1.0.1",
-    "methods": "~1.1.2",
-    "on-finished": "~2.3.0",
-    "parseurl": "~1.3.2",
-    "path-to-regexp": "0.1.7",
-    "proxy-addr": "~2.0.4",
-    "qs": "6.5.2",
-    "range-parser": "~1.2.0",
-    "safe-buffer": "5.1.2",
-    "send": "0.16.2",
-    "serve-static": "1.13.2",
-    "setprototypeof": "1.1.0",
-    "statuses": "~1.4.0",
-    "type-is": "~1.6.16",
-    "utils-merge": "1.0.1",
-    "vary": "~1.1.2"
+    "accepts": "^2.0.0",
+    "body-parser": "^2.2.0",
+    "content-disposition": "^1.0.0",
+    "content-type": "^1.0.5",
+    "cookie": "^0.7.1",
+    "cookie-signature": "^1.2.1",
+    "debug": "^4.4.0",
+    "encodeurl": "^2.0.0",
+    "escape-html": "^1.0.3",
+    "etag": "^1.8.1",
+    "finalhandler": "^2.1.0",
+    "fresh": "^2.0.0",
+    "http-errors": "^2.0.0",
+    "merge-descriptors": "^2.0.0",
+    "mime-types": "^3.0.0",
+    "on-finished": "^2.4.1",
+    "once": "^1.4.0",
+    "parseurl": "^1.3.3",
+    "proxy-addr": "^2.0.7",
+    "qs": "^6.14.0",
+    "range-parser": "^1.2.1",
+    "router": "^2.2.0",
+    "send": "^1.1.0",
+    "serve-static": "^2.2.0",
+    "statuses": "^2.0.1",
+    "type-is": "^2.0.1",
+    "vary": "^1.1.2"
   },
-  "deprecated": false,
-  "description": "Fast, unopinionated, minimalist web framework",
   "devDependencies": {
     "after": "0.8.2",
-    "connect-redis": "3.4.0",
-    "cookie-parser": "~1.4.3",
-    "cookie-session": "1.3.2",
-    "ejs": "2.6.1",
-    "eslint": "2.13.1",
-    "express-session": "1.15.6",
-    "hbs": "4.0.1",
-    "istanbul": "0.4.5",
-    "marked": "0.5.1",
+    "connect-redis": "^8.0.1",
+    "cookie-parser": "1.4.7",
+    "cookie-session": "2.1.0",
+    "ejs": "^3.1.10",
+    "eslint": "8.47.0",
+    "express-session": "^1.18.1",
+    "hbs": "4.2.0",
+    "marked": "^15.0.3",
     "method-override": "3.0.0",
-    "mocha": "5.2.0",
-    "morgan": "1.9.1",
-    "multiparty": "4.2.1",
+    "mocha": "^10.7.3",
+    "morgan": "1.10.0",
+    "nyc": "^17.1.0",
     "pbkdf2-password": "1.2.1",
-    "should": "13.2.3",
-    "supertest": "3.3.0",
+    "supertest": "^6.3.0",
     "vhost": "~3.0.2"
   },
   "engines": {
-    "node": ">= 0.10.0"
+    "node": ">= 18"
   },
   "files": [
     "LICENSE",
