@@ -1,53 +1,29 @@
 {
-  "_from": "forwarded@~0.1.2",
-  "_id": "forwarded@0.1.2",
-  "_inBundle": false,
-  "_integrity": "sha1-mMI9qxF1ZXuMBXPozszZGw/xjIQ=",
-  "_location": "/forwarded",
-  "_phantomChildren": {},
-  "_requested": {
-    "type": "range",
-    "registry": true,
-    "raw": "forwarded@~0.1.2",
-    "name": "forwarded",
-    "escapedName": "forwarded",
-    "rawSpec": "~0.1.2",
-    "saveSpec": null,
-    "fetchSpec": "~0.1.2"
-  },
-  "_requiredBy": [
-    "/proxy-addr"
-  ],
-  "_resolved": "https://registry.npmjs.org/forwarded/-/forwarded-0.1.2.tgz",
-  "_shasum": "98c23dab1175657b8c0573e8ceccd91b0ff18c84",
-  "_spec": "forwarded@~0.1.2",
-  "_where": "/Users/didin/NodeApps/node-graphql/node_modules/proxy-addr",
-  "bugs": {
-    "url": "https://github.com/jshttp/forwarded/issues"
-  },
-  "bundleDependencies": false,
-  "contributors": [
-    {
-      "name": "Douglas Christopher Wilson",
-      "email": "doug@somethingdoug.com"
-    }
-  ],
-  "deprecated": false,
+  "name": "forwarded",
   "description": "Parse HTTP X-Forwarded-For header",
+  "version": "0.2.0",
+  "contributors": [
+    "Douglas Christopher Wilson <doug@somethingdoug.com>"
+  ],
+  "license": "MIT",
+  "keywords": [
+    "x-forwarded-for",
+    "http",
+    "req"
+  ],
+  "repository": "jshttp/forwarded",
   "devDependencies": {
     "beautify-benchmark": "0.2.4",
     "benchmark": "2.1.4",
-    "eslint": "3.19.0",
-    "eslint-config-standard": "10.2.1",
-    "eslint-plugin-import": "2.7.0",
-    "eslint-plugin-node": "5.1.1",
-    "eslint-plugin-promise": "3.5.0",
-    "eslint-plugin-standard": "3.0.1",
-    "istanbul": "0.4.5",
-    "mocha": "1.21.5"
-  },
-  "engines": {
-    "node": ">= 0.6"
+    "deep-equal": "1.0.1",
+    "eslint": "7.27.0",
+    "eslint-config-standard": "14.1.1",
+    "eslint-plugin-import": "2.23.4",
+    "eslint-plugin-node": "11.1.0",
+    "eslint-plugin-promise": "4.3.1",
+    "eslint-plugin-standard": "4.1.0",
+    "mocha": "8.4.0",
+    "nyc": "15.1.0"
   },
   "files": [
     "LICENSE",
@@ -55,24 +31,15 @@
     "README.md",
     "index.js"
   ],
-  "homepage": "https://github.com/jshttp/forwarded#readme",
-  "keywords": [
-    "x-forwarded-for",
-    "http",
-    "req"
-  ],
-  "license": "MIT",
-  "name": "forwarded",
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/jshttp/forwarded.git"
+  "engines": {
+    "node": ">= 0.6"
   },
   "scripts": {
     "bench": "node benchmark/index.js",
     "lint": "eslint .",
     "test": "mocha --reporter spec --bail --check-leaks test/",
-    "test-cov": "istanbul cover node_modules/mocha/bin/_mocha -- --reporter dot --check-leaks test/",
-    "test-travis": "istanbul cover node_modules/mocha/bin/_mocha --report lcovonly -- --reporter spec --check-leaks test/"
-  },
-  "version": "0.1.2"
+    "test-ci": "nyc --reporter=lcov --reporter=text npm test",
+    "test-cov": "nyc --reporter=html --reporter=text npm test",
+    "version": "node scripts/version-history.js && git add HISTORY.md"
+  }
 }
